feat(checkout): CHECKOUT-9477 refactor Shipping component
