+++ packages/core/src/app/shipping/hooks/useShipping.ts
@@ -1,0 +1,130 @@
+import { createSelector } from "reselect";
+
+import { shouldUseStripeLinkByMinimumAmount } from "@bigcommerce/checkout/instrument-utils";
+import { type CheckoutContextProps, PaymentMethodId, useCheckout } from "@bigcommerce/checkout/payment-integration-api";
+
+import { EMPTY_ARRAY } from "../../common/utility";
+import getProviderWithCustomCheckout from "../../payment/getProviderWithCustomCheckout";
+import getShippableItemsCount from "../getShippableItemsCount";
+import getShippingMethodId from "../getShippingMethodId";
+import hasPromotionalItems from "../hasPromotionalItems";
+
+
+const deleteConsignmentsSelector = createSelector(
+    ({ checkoutService: { deleteConsignment } }: CheckoutContextProps) => deleteConsignment,
+    ({ checkoutState: { data } }: CheckoutContextProps) => data.getConsignments(),
+    (deleteConsignment, consignments) => async () => {
+        if (!consignments || !consignments.length) {
+            return;
+        }
+
+        const [{ data }] = await Promise.all(consignments.map(({ id }) => deleteConsignment(id)));
+
+        return data.getShippingAddress();
+    },
+);
+
+export const useShipping = () => {
+    const { checkoutState, checkoutService } = useCheckout();
+
+    const {
+        data: {
+            getCart,
+            getCheckout,
+            getConfig,
+            getCustomer,
+            getConsignments,
+            getShippingAddress,
+            getBillingAddress,
+            getShippingAddressFields,
+            getShippingCountries,
+        },
+        statuses: {
+            isShippingStepPending,
+            isSelectingShippingOption,
+            isLoadingShippingOptions,
+            isUpdatingConsignment,
+            isCreatingConsignments,
+            isCreatingCustomerAddress,
+            isLoadingShippingCountries,
+            isUpdatingBillingAddress,
+            isUpdatingCheckout,
+            isDeletingConsignment,
+            isLoadingCheckout,
+        },
+    } = checkoutState;
+
+    const checkout = getCheckout();
+    const config = getConfig();
+    const consignments = getConsignments() || [];
+    const customer = getCustomer();
+    const cart = getCart();
+
+    if (!checkout || !config || !customer || !cart) {
+        throw new Error('Unable to access checkout data');
+    }
+
+    const {
+        checkoutSettings: {
+            enableOrderComments,
+            hasMultiShippingEnabled,
+        },
+    } = config;
+
+    const methodId = getShippingMethodId(checkout, config);
+    const isLoading =
+        isLoadingShippingOptions() ||
+        isSelectingShippingOption() ||
+        isUpdatingConsignment() ||
+        isCreatingConsignments() ||
+        isUpdatingBillingAddress() ||
+        isUpdatingCheckout() ||
+        isCreatingCustomerAddress() ||
+        isDeletingConsignment() ||
+        isLoadingCheckout();
+
+    const shippableItemsCount = getShippableItemsCount(cart);
+    const shouldShowMultiShipping =
+        hasMultiShippingEnabled && !methodId && shippableItemsCount > 1;
+
+    const shippingAddress =
+        !shouldShowMultiShipping && consignments.length > 1 ? undefined : getShippingAddress();
+
+    const providerWithCustomCheckout = getProviderWithCustomCheckout(
+        config.checkoutSettings.providerWithCustomCheckout,
+    );
+
+    return {
+        assignItem: checkoutService.assignItemsToAddress,
+        billingAddress: getBillingAddress(),
+        cart,
+        cartHasPromotionalItems: hasPromotionalItems(cart),
+        consignments,
+        countries: getShippingCountries() || EMPTY_ARRAY,
+        customer,
+        customerMessage: checkout.customerMessage,
+        createCustomerAddress: checkoutService.createCustomerAddress,
+        deinitializeShippingMethod: checkoutService.deinitializeShipping,
+        deleteConsignments: deleteConsignmentsSelector({ checkoutService, checkoutState }),
+        getFields: getShippingAddressFields,
+        initializeShippingMethod: checkoutService.initializeShipping,
+        isGuest: customer.isGuest,
+        isInitializing: isLoadingShippingCountries() || isLoadingShippingOptions(),
+        isLoading,
+        isShippingStepPending: isShippingStepPending(),
+        loadShippingAddressFields: checkoutService.loadShippingAddressFields,
+        loadBillingAddressFields: checkoutService.loadBillingAddressFields,
+        loadShippingOptions: checkoutService.loadShippingOptions,
+        methodId,
+        providerWithCustomCheckout,
+        shippingAddress,
+        shouldShowMultiShipping,
+        shouldShowOrderComments: enableOrderComments,
+        signOut: checkoutService.signOutCustomer,
+        unassignItem: checkoutService.unassignItemsToAddress,
+        updateBillingAddress: checkoutService.updateBillingAddress,
+        updateCheckout: checkoutService.updateCheckout,
+        updateShippingAddress: checkoutService.updateShippingAddress,
+        shouldRenderStripeForm: providerWithCustomCheckout === PaymentMethodId.StripeUPE && shouldUseStripeLinkByMinimumAmount(cart),
+    };
+}
