+++ packages/core/src/app/shipping/stripeUPE/StripeShipping.test.tsx
@@ -4,14 +4,17 @@ import {
 } from '@bigcommerce/checkout-sdk';
 import React from 'react';
 
+import { type AnalyticsEvents, AnalyticsProviderMock } from '@bigcommerce/checkout/analytics';
 import { ExtensionProvider } from '@bigcommerce/checkout/checkout-extension';
 import { LocaleProvider } from '@bigcommerce/checkout/locale';
 import { CheckoutProvider } from '@bigcommerce/checkout/payment-integration-api';
 import { render } from '@bigcommerce/checkout/test-utils';
 
+import { getCart } from '../../cart/carts.mock';
 import { getCheckout } from '../../checkout/checkouts.mock';
 import CheckoutStepType from '../../checkout/CheckoutStepType';
 import ConsoleErrorLogger from '../../common/error/ConsoleErrorLogger';
+import { getStoreConfig } from '../../config/config.mock';
 import { getCustomer } from '../../customer/customers.mock';
 import { getShippingAddress } from '../shipping-addresses.mock';
 
@@ -22,6 +25,7 @@ describe('Stripe Shipping Component', () => {
     const errorLogger = new ConsoleErrorLogger();
     const initialize = jest.fn();
     let checkoutState: CheckoutSelectors;
+    let analyticsTracker: Partial<AnalyticsEvents>;
 
    checkoutService.initializeShipping = initialize;
 
@@ -37,35 +41,36 @@ describe('Stripe Shipping Component', () => {
         isBillingSameAsShipping: false,
         isInitialValueLoaded: false,
         isMultiShippingMode: false,
-        shippingAddress: getShippingAddress(),
-        shouldShowOrderComments: true,
         cartHasChanged: false,
         isLoading: false,
-        isShippingStepPending: false,
-        isInitializing: true,
-        isShippingMethodLoading: true,
-        shouldShowMultiShipping: false,
-        loadShippingAddressFields: jest.fn(),
-        loadShippingOptions: jest.fn(),
         onMultiShippingChange: jest.fn(),
         onSubmit: jest.fn(),
         onUnhandledError: jest.fn(),
-        updateAddress: jest.fn(),
     };
 
     beforeEach(() => {
       checkoutState = checkoutService.getState();
+      analyticsTracker = {
+          checkoutBegin: jest.fn(),
+          trackStepViewed: jest.fn(),
+          trackStepCompleted: jest.fn(),
+          exitCheckout: jest.fn(),
+      };
       jest.spyOn(checkoutState.data, 'getCustomer').mockReturnValue(getCustomer());
       jest.spyOn(checkoutState.data, 'getCheckout').mockReturnValue(getCheckout());
+      jest.spyOn(checkoutState.data, 'getConfig').mockReturnValue(getStoreConfig());
+      jest.spyOn(checkoutState.data, 'getCart').mockReturnValue(getCart());
     });
 
     it('loads shipping data  when component is mounted', async () => {
       const { container } = render(
         <CheckoutProvider checkoutService={checkoutService}>
           <LocaleProvider checkoutService={checkoutService}>
-            <ExtensionProvider checkoutService={checkoutService} errorLogger={errorLogger}>
-              <StripeShipping {...defaultProps} />
-            </ExtensionProvider>
+            <AnalyticsProviderMock analyticsTracker={analyticsTracker}>
+              <ExtensionProvider checkoutService={checkoutService} errorLogger={errorLogger}>
+                <StripeShipping {...defaultProps} />
+              </ExtensionProvider>
+            </AnalyticsProviderMock>
           </LocaleProvider>
         </CheckoutProvider>,
       );
